refactor: 패키지 구조 수정
- Reorganized package structure to include `company` namespace.
- Updated `Company` entity to extend `BaseAuditEntity` and added schema to table annotation.
- Modified logging in `KaKaoAddress` for geolocation details.
- Adjusted `CompanyLocation.of` method parameter order for consistency.
