+++ company-service/src/main/java/com/winner/client/companyservice/company/infrastructure/api/KaKaoAddress.java
@@ -1,7 +1,7 @@
-package com.winner.client.companyservice.infrastructure.api;
+package com.winner.client.companyservice.company.infrastructure.api;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import com.winner.client.companyservice.infrastructure.service.GeocodingService;
+import com.winner.client.companyservice.company.infrastructure.service.GeocodingService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
@@ -39,6 +39,9 @@ public class KaKaoAddress implements GeocodingService {
         JsonNode node = jsonNode.get("documents").get(0);
           result[0] = node.get("x").asDouble();
           result[1] = node.get("y").asDouble();
+
+          log.info("경도 : {}",result[0]);
+          log.info("위도 : {}",result[1]);
 
     }catch(RuntimeException e){
       log.error("카카오 요청 오류 : {}", e.getMessage());
