+++ company-service/src/main/java/com/winner/client/companyservice/company/presentation/dto/CompanyRequestDto.java
@@ -1,7 +1,7 @@
-package com.winner.client.companyservice.presentation.dto;
+package com.winner.client.companyservice.company.presentation.dto;
 
-import com.winner.client.companyservice.application.dto.CompanyServiceDto;
-import com.winner.client.companyservice.domain.vo.Type;
+import com.winner.client.companyservice.company.application.dto.CompanyServiceDto;
+import com.winner.client.companyservice.company.entity.vo.Type;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import java.util.UUID;
